EDM-1476: add Events validation TCs
